% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mass.R
\name{mass}
\alias{mass}
\title{Run MASS}
\usage{
mass(file, outdir = dirname(file), nregs = NULL,
  template_directory = system.file(file.path("mass/share/data/Templates",
  ifelse(cerebellum, "WithCerebellum", "WithoutCerebellum")), package =
  "massr"), cerebellum = TRUE)
}
\arguments{
\item{file}{Filename (or nifti) to match to target}

\item{outdir}{Output directory for results}

\item{nregs}{Number of registrations to do (default is 6)}

\item{template_directory}{Directory}

\item{retimg}{return nifti object versus output image}
}
\value{
Result of \code{\link{system}} command
}
\description{
Runs MASS on source and target
}
\examples{
\dontrun{
require(fslr)
temp = file.path(fsldir(), "data", "standard", "MNI152_T1_1mm.nii.gz")
res = mass(file = temp)
}
}

