% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mass_cmd_maker.R
\name{mass_cmd_maker}
\alias{mass_cmd_maker}
\title{Create MASS command}
\usage{
mass_cmd_maker(cmd = NULL, args = NULL)
}
\arguments{
\item{cmd}{Command to be executed}

\item{args}{Arguments to be added to the command.}
}
\value{
character string of command
}
\description{
Create header for MASS command
}

